CBufferPayload moved to separate header ecal_buffer_payload.h
