+++ ecal/core/include/ecal/ecal_payload.h
@@ -25,7 +25,6 @@
 #pragma once
 
 #include <cstddef>
-#include <cstring>
 
 namespace eCAL
 {
@@ -38,38 +37,17 @@ namespace eCAL
   public:
     virtual ~CPayload() = default;
 
+    // the provisioned memory is uninitialized ->
     // perform a full write operation
-    // the provisioned memory is uninitialized
     virtual bool WriteComplete (void* buf_, size_t len_) = 0;
 
-    // the provisioned memory is initialized and contains the data from the last write operation,
-    // so you can perform a partial write operation or just modify a few bytes here
+    // the provisioned memory is initialized and contains the data from the last write operation ->
+    // perform a partial write operation or just modify a few bytes here
     //
     // by default this operation will just call `WriteComplete`
     virtual bool WritePartial  (void* buf_, size_t len_) { return WriteComplete(buf_, len_); };
 
     // provide the size of the required memory (eCAL needs to allocate for you).
     virtual size_t GetSize() = 0;
-  };
-
-  // specific payload class to wrap classic (void*, size_t) interface
-  class CBufferPayload : public CPayload
-  {
-  public:
-    CBufferPayload(const void* const buf_, size_t len_) : m_buf(buf_), m_buf_len(len_) {};
-
-    // make a dump memory copy
-    bool WriteComplete (void* buf_, size_t len_) {
-      if (len_ < m_buf_len) return false;
-      memcpy(buf_, m_buf, m_buf_len);
-      return true;
-    }
-
-    // size of the memory that needs to be copied
-    size_t GetSize() { return m_buf_len; };
-  
-  private:
-    const void* const m_buf     = nullptr;
-    size_t            m_buf_len = 0;
   };
 };
