+++ ecal/core/src/pubsub/ecal_publisher.cpp
@@ -26,6 +26,7 @@
 #include <ecal/ecal_config.h>
 
 #include "ecal_def.h"
+#include "ecal_buffer_payload.h"
 #include "ecal_config_reader_hlp.h"
 #include "ecal_globals.h"
 #include "ecal_pubgate.h"
